Add front-, main-, and backmatter.
If the commands are not defined by the document class, e.g. in article
or report, one can use them there, too.
The numbering of pages in the front matter is done by small roman
numbers, the other two have arabic numbers. In the back matter the
chapter mark is on every page and not only on every other.