Get rid of incompatible `titlesec` package
The `titlesec` package does some very crazy things that are incompatible
with KOMAScript. When using the two together, KOMAScript reports this
issue and suggests to not using `titlesec`.

From now on, the layout of the headlines is done by KOMAScript, tweaking
it to almost the same look as with `titlesec`.

Add the option `chapterprefix=true` when loading a document class from
KOMAScript to support the feature of separating the chapter number into
the margin paragraph.